+++ packages/react-ui/storybook/stories/widgetsUI/InputFile.stories.js
@@ -50,6 +50,10 @@ const options = {
       source: {
         type: 'auto'
       }
+    },
+    design: {
+      type: 'figma',
+      url: 'https://www.figma.com/file/nmaoLeo69xBJCHm9nc6lEV/CARTO-Components-1.0?node-id=1159%3A24236'
     }
   }
 };